+++ rich/console.py
@@ -10,6 +10,7 @@ from dataclasses import dataclass, field, replace
 from datetime import datetime
 from functools import wraps
 from getpass import getpass
+from itertools import islice
 from time import monotonic
 from typing import (
     IO,
@@ -22,6 +23,7 @@ from typing import (
     NamedTuple,
     Optional,
     TextIO,
+    Tuple,
     Union,
     cast,
 )
@@ -57,6 +59,13 @@ WINDOWS = platform.system() == "Windows"
 HighlighterType = Callable[[Union[str, "Text"]], "Text"]
 JustifyMethod = Literal["default", "left", "center", "right", "full"]
 OverflowMethod = Literal["fold", "crop", "ellipsis", "ignore"]
+
+
+class NoChange:
+    pass
+
+
+NO_CHANGE = NoChange()
 
 
 CONSOLE_HTML_FORMAT = """\
@@ -105,6 +114,8 @@ class ConsoleOptions:
     """Disable wrapping for text."""
     highlight: Optional[bool] = None
     """Highlight override for render_str."""
+    height: Optional[int] = None
+    """Height available, or None for no height limit."""
 
     @property
     def ascii_only(self) -> bool:
@@ -113,30 +124,45 @@ class ConsoleOptions:
 
     def update(
         self,
-        width: int = None,
-        min_width: int = None,
-        max_width: int = None,
-        justify: JustifyMethod = None,
-        overflow: OverflowMethod = None,
-        no_wrap: bool = None,
-        highlight: bool = None,
+        width: Union[int, NoChange] = NO_CHANGE,
+        min_width: Union[int, NoChange] = NO_CHANGE,
+        max_width: Union[int, NoChange] = NO_CHANGE,
+        justify: Union[Optional[JustifyMethod], NoChange] = NO_CHANGE,
+        overflow: Union[Optional[OverflowMethod], NoChange] = NO_CHANGE,
+        no_wrap: Union[Optional[bool], NoChange] = NO_CHANGE,
+        highlight: Union[Optional[bool], NoChange] = NO_CHANGE,
+        height: Union[Optional[int], NoChange] = NO_CHANGE,
     ) -> "ConsoleOptions":
         """Update values, return a copy."""
         options = replace(self)
-        if width is not None:
+        if not isinstance(width, NoChange):
             options.min_width = options.max_width = width
-        if min_width is not None:
+        if not isinstance(min_width, NoChange):
             options.min_width = min_width
-        if max_width is not None:
+        if not isinstance(max_width, NoChange):
             options.max_width = max_width
-        if justify is not None:
+        if not isinstance(justify, NoChange):
             options.justify = justify
-        if overflow is not None:
+        if not isinstance(overflow, NoChange):
             options.overflow = overflow
-        if no_wrap is not None:
+        if not isinstance(no_wrap, NoChange):
             options.no_wrap = no_wrap
-        if highlight is not None:
+        if not isinstance(highlight, NoChange):
             options.highlight = highlight
+        if not isinstance(height, NoChange):
+            options.height = height
+        return options
+
+    def update_width(self, width: int) -> "ConsoleOptions":
+        """Update just the width, return a copy.
+
+        Args:
+            width (int): New width (sets both min_width and max_width)
+
+        Returns:
+            ~ConsoleOptions: New console options instance
+        """
+        options = replace(self, min_width=width, max_width=width)
         return options
 
 
@@ -253,17 +279,22 @@ class PagerContext:
 class ScreenContext:
     """A context manager that enables an alternative screen. See :meth:`~rich.console.Console.screen` for usage."""
 
-    def __init__(self, console: "Console") -> None:
+    def __init__(self, console: "Console", hide_cursor: bool) -> None:
         self.console = console
+        self.hide_cursor = hide_cursor
         self._changed = False
 
     def __enter__(self) -> "ScreenContext":
         self._changed = self.console.set_alt_screen(True)
+        if self._changed and self.hide_cursor:
+            self.console.show_cursor(False)
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb) -> None:
         if self._changed:
             self.console.set_alt_screen(False)
+            if self.hide_cursor:
+                self.console.show_cursor(True)
 
 
 class RenderGroup:
@@ -787,6 +818,16 @@ class Console:
         width, _ = self.size
         return width
 
+    @property
+    def height(self) -> int:
+        """Get the height of the console.
+
+        Returns:
+            int: The height (in lines) of the console.
+        """
+        _, height = self.size
+        return height
+
     def bell(self) -> None:
         """Play a 'bell' sound (if supported by the terminal)."""
         self.control("\x07")
@@ -910,18 +951,19 @@ class Console:
             bool: True if the control codes were written.
 
         """
+        changed = False
         if self.is_terminal and not self.legacy_windows:
             self.control("\033[?1049h\033[H" if enable else "\033[?1049l")
-            return True
-        return False
+            changed = True
+        return changed
 
-    def screen(self) -> "ScreenContext":
+    def screen(self, hide_cursor: bool = True) -> "ScreenContext":
         """Context manager to enable and disable 'alternative screen' mode.
 
         Returns:
             ~ScreenContext: Context which enables alternate screen on enter, and disables it on exit.
         """
-        return ScreenContext(self)
+        return ScreenContext(self, hide_cursor=hide_cursor)
 
     def render(
         self, renderable: RenderableType, options: ConsoleOptions = None
@@ -990,6 +1032,7 @@ class Console:
             options (Optional[ConsoleOptions], optional): Console options, or None to use self.options. Default to ``None``.
             style (Style, optional): Optional style to apply to renderables. Defaults to ``None``.
             pad (bool, optional): Pad lines shorter than render width. Defaults to ``True``.
+            range (Optional[Tuple[int, int]], optional): Range of lines to render, or ``None`` for all line. Defaults to ``None``
 
         Returns:
             List[List[Segment]]: A list of lines, where a line is a list of Segment objects.
@@ -1003,6 +1046,10 @@ class Console:
                 _rendered, render_options.max_width, include_new_lines=False, pad=pad
             )
         )
+        if render_options.height is not None:
+            lines = Segment.set_shape(
+                lines, render_options.max_width, render_options.height, style=style
+            )
         return lines
 
     def render_str(
@@ -1244,6 +1291,7 @@ class Console:
         markup: bool = None,
         highlight: bool = None,
         width: int = None,
+        height: int = None,
         crop: bool = True,
         soft_wrap: bool = None,
     ) -> None:
@@ -1293,9 +1341,11 @@ class Console:
             render_options = self.options.update(
                 justify="default",
                 overflow=overflow,
-                width=min(width, self.width) if width else None,
+                width=min(width, self.width) if width else NO_CHANGE,
+                height=height,
                 no_wrap=no_wrap,
             )
+
             new_segments: List[Segment] = []
             extend = new_segments.extend
             render = self.render