+++ tests/test_traceback.py
@@ -81,6 +81,19 @@ def test_print_exception():
     assert "ZeroDivisionError" in exception_text
 
 
+def test_print_exception_locals():
+    console = Console(width=100, file=io.StringIO())
+    try:
+        1 / 0
+    except Exception:
+        console.print_exception(show_locals=True)
+    exception_text = console.file.getvalue()
+    print(exception_text)
+    assert "ZeroDivisionError" in exception_text
+    assert "locals" in exception_text
+    assert "console = <console width=100 None>" in exception_text
+
+
 def test_syntax_error():
     console = Console(width=100, file=io.StringIO())
     try:
@@ -179,6 +192,13 @@ def test_broken_str():
     result = console.file.getvalue()
     print(result)
     assert "<exception str() failed>" in result
+
+
+def test_guess_lexer():
+    assert Traceback._guess_lexer("foo.py", "code") == "python"
+    code_python = "#! usr/bin/env python\nimport this"
+    assert Traceback._guess_lexer("foo", code_python) == "python"
+    assert Traceback._guess_lexer("foo", "foo\nbnar") == "text"
 
 
 if __name__ == "__main__":  # pragma: no cover